pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "newsletter-generator"
version = "0.1.0"
description = "내부 연구원을 위한 뉴스레터 생성 CLI 도구"
readme = "README.md"
authors = [
    {name = "Newsletter Generator Team", email = "example@corp.com"},
]
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.5.0",
    "requests>=2.31.0",
    "google-generativeai>=0.3.0",
    "jinja2>=3.1.2",
    "python-dotenv>=1.0.0",
    "sendgrid>=6.10.0",
    "google-api-python-client>=2.100.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=1.0.0",
    "pydantic>=2.0.0",
    "sqlite-utils>=3.35.0",
]

[project.scripts]
newsletter = "newsletter.cli:app"

[tool.setuptools]
packages = ["newsletter"]
